+++ apps/web/src/components/purchase-form/cards/add-card.tsx
@@ -6,34 +6,20 @@ import AddMethodsForm from './sections/add-methods-form'
 import AddMethodsSuccess from './sections/add-methods-success'
 
 import Loading from '@/components/loading/loading'
-import { SelectOption } from '@/components/select-input/select-input'
-import { FormValidation } from '@/contexts/payment-context'
+import { usePaymentContext } from '@/contexts/payment-context'
 
 export interface PaymentMethodsFormProps {
-  countries: SelectOption[]
-  formErrors?: FormValidation
-  handleRetry: () => void
-  loadingText: string
   onSubmit(event: FormEvent<HTMLFormElement>): void
-  status?: CheckoutStatus
 }
 
 export default function PaymentMethodsForm({
-  countries,
-  handleRetry,
-  formErrors,
-  loadingText,
   onSubmit,
-  status,
 }: PaymentMethodsFormProps) {
+  const { handleRetry, loadingText, status } = usePaymentContext()
   return (
     <section className="pt-5">
       <div className={status === CheckoutStatus.form ? '' : 'hidden'}>
-        <AddMethodsForm
-          countries={countries}
-          formErrors={formErrors}
-          onSubmit={onSubmit}
-        />
+        <AddMethodsForm onSubmit={onSubmit} />
       </div>
 
       {status === CheckoutStatus.loading && (
