+++ apps/web/src/components/purchase-form/cards/sections/add-methods-form.tsx
@@ -9,29 +9,26 @@ import Heading from '@/components/heading'
 import BillingAddress from '@/components/purchase-form/shared/billing-address'
 import CardDetails from '@/components/purchase-form/shared/card-details'
 import FullName from '@/components/purchase-form/shared/full-name'
-import { FormValidation, getError } from '@/contexts/payment-context'
+import { usePaymentContext } from '@/contexts/payment-context'
 
 export interface AddMethodsFormProps {
-  formErrors?: FormValidation
   onSubmit(event: FormEvent<HTMLFormElement>): void
 }
 
-export default function AddMethodsForm({
-  formErrors,
-  onSubmit,
-}: AddMethodsFormProps) {
+export default function AddMethodsForm({ onSubmit }: AddMethodsFormProps) {
   const { t } = useTranslation()
+  const { getError } = usePaymentContext()
   return (
     <form className={css.form} onSubmit={onSubmit}>
-      {getError('expirationDate', formErrors) ? (
+      {getError('expirationDate') ? (
         <AlertMessage
           className={css.notification}
-          content={getError('expirationDate', formErrors)}
+          content={getError('expirationDate')}
           variant="red"
         />
       ) : null}
       <Heading level={2}>{t('forms:sections.Credit Card')}</Heading>
-      <FullName formErrors={{ fullName: getError('fullName', formErrors) }} />
+      <FullName formErrors={{ fullName: getError('fullName') }} />
       <CardDetails />
       <BillingAddress />
       {/* Submit */}
